Update documentation. Remove unused code. Split API extras/examples to separate file for clarity.
