+++ prez/routers/ogc_router.py
@@ -1,6 +1,3 @@
-import json
-from pathlib import Path
-
 from fastapi import APIRouter, Depends
 from sparql_grammar_pydantic import ConstructQuery
 
@@ -18,6 +15,7 @@ from prez.dependencies import (
 from prez.models.query_params import QueryParams
 from prez.reference_data.prez_ns import EP, ONT, OGCE
 from prez.repositories import Repo
+from prez.routers.api_extras_examples import responses, cql_examples, openapi_extras
 from prez.services.connegp_service import NegotiatedPMTs
 from prez.services.listings import listing_function
 from prez.services.objects import object_function
@@ -26,124 +24,6 @@ from prez.services.query_generation.cql import CQLParser
 from prez.services.query_generation.shacl import NodeShape
 
 router = APIRouter(tags=["ogcprez"])
-
-responses_json = Path(__file__).parent / "rdf_response_examples.json"
-responses = json.loads(responses_json.read_text())
-# responses = {
-#     200: {
-#         "content": {
-#             "application/ld+json": {
-#                 "example": {
-#                     "@id": "https://example.com/item/1",
-#                     "https://example.com/property": "value"
-#                 }
-#             },
-#             "application/anot+ld+json": {
-#                 "example": {
-#                     "@context": {"prez": "https://prez.dev/"},
-#                     "@id": "https://example.com/item/1",
-#                     "https://example.com/property": "value",
-#                     "prez:label": "Item One"
-#                 }
-#             },
-#             "application/rdf+xml": {
-#                 "example": "<?xml version=\"1.0\"?>\n<rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n    <rdf:Description rdf:about=\"https://example.com/item/1\">\n        <ns1:property xmlns:ns1=\"https://example.com/\">value</ns1:property>\n    </rdf:Description>\n</rdf:RDF>"
-#             },
-#             "text/anot+turtle": {
-#                 "example": """
-#                     @prefix prez: <https://prez.dev/> .
-#
-#                     <https://example.com/item/1>
-#                         <https://example.com/property> "value" ;
-#                         prez:label "Item One" .
-#                 """
-#             },
-#             "text/turtle": {
-#                 "example": """
-#                     <https://example.com/item/1> <https://example.com/property> "value" .
-#                 """
-#             },
-#         }
-#     }
-# }
-
-
-# Path to the directory containing JSON files
-cql_json_examples_dir = Path(__file__).parent.parent / "examples/cql"
-
-# Dictionary comprehension to create examples
-cql_examples = {
-    file.stem: {
-        "summary": file.stem,
-        "value": json.loads(file.read_text())
-    }
-    for file in cql_json_examples_dir.glob("*.json")
-}
-
-
-def create_path_param(name: str, description: str, example: str):
-    return {
-        "in": "path",
-        "name": name,
-        "required": True,
-        "schema": {
-            "type": "string",
-            "example": example,
-        },
-        "description": description,
-    }
-
-
-path_parameters = {
-    "collection-listing": [
-        create_path_param("catalogId", "Curie of the Catalog ID.", "bblck-ctlg:bblocks")
-    ],
-    "item-listing": [
-        create_path_param(
-            "catalogId", "Curie of the Catalog ID.", "bblck-ctlg:bblocks"
-        ),
-        create_path_param(
-            "collectionId", "Curie of the Collection ID.", "bblck-vcbs:api"
-        ),
-    ],
-    "top-concepts": [
-        create_path_param("parent_curie", "Parent CURIE.", "exm:SchemingConceptScheme")
-    ],
-    "narrowers": [
-        create_path_param("parent_curie", "Parent CURIE.", "exm:TopLevelConcept")
-    ],
-    "profile-object": [
-        create_path_param("profile_curie", "Profile CURIE.", "prez:OGCItemProfile")
-    ],
-    "catalog-object": [
-        create_path_param("catalogId", "Catalog ID.", "bblck-ctlg:bblocks")
-    ],
-    "collection-object": [
-        create_path_param("catalogId", "Catalog ID.", "bblck-ctlg:bblocks"),
-        create_path_param("collectionId", "Collection ID.", "bblck-vcbs:api"),
-    ],
-    "item-object": [
-        create_path_param("catalogId", "Catalog ID.", "bblck-ctlg:bblocks"),
-        create_path_param("collectionId", "Collection ID.", "bblck-vcbs:api"),
-        create_path_param("itemId", "Item ID.", "bblcks:ogc.unstable.sosa"),
-    ],
-}
-
-openapi_extras = {
-    name: {"parameters": params} for name, params in path_parameters.items()
-}
-
-
-########################################################################################################################
-# Listing endpoints
-
-# 1: /profiles
-# 2: /catalogs
-# 3: /catalogs/{catalogId}/collections
-# 4: /catalogs/{catalogId}/collections/{collectionId}/items
-# 5: /concept-hierarchy/{parent_uri}/top-concepts
-# 6: /concept-hierarchy/{parent_uri}/narrowers
-########################################################################################################################
 
 
 @router.get(
